Implement exception

+++ src/main/java/no/api/web/v1/controller/BasketController.java
@@ -1,5 +1,6 @@
 package no.api.web.v1.controller;
 
+import no.api.exception.ResourceNotFoundException;
 import no.api.model.BasketModel;
 import no.api.model.BasketProductModel;
 import no.api.service.BasketProductService;
@@ -11,11 +12,10 @@ import no.api.web.v1.mapper.BasketResponseMapper;
 import no.api.web.v1.transport.BasketProductTransport;
 import no.api.web.v1.transport.BasketTransport;
 import no.api.web.v1.transport.ResponseTransport;
-import org.jsondoc.core.annotation.Api;
-import org.jsondoc.core.annotation.ApiBodyObject;
-import org.jsondoc.core.annotation.ApiMethod;
-import org.jsondoc.core.annotation.ApiPathParam;
+import org.jsondoc.core.annotation.*;
 import org.jsondoc.core.pojo.ApiVerb;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -29,6 +29,12 @@ import java.util.List;
 @RestController
 @RequestMapping(value = "/api/b2b/v1/baskets", produces = MediaType.APPLICATION_JSON_VALUE)
 public class BasketController {
+    private static final Logger LOGGER = LoggerFactory.getLogger(BasketController.class);
+    private static final String BAD_REQUEST_STATUS_CODE = "400";
+    private static final String NOT_FOUND_STATUS_CODE = "404";
+    private static final String UNSUPPORT_MEDIA_TYPE_STATUS_CODE = "415";
+    private static final String UNPROCESS_ENTITY__STATUS_CODE = "422";
+    private static final String SERVER_ERROR_STATUS_CODE = "500";
 
     @Autowired
     private BasketService basketService;
@@ -41,8 +47,15 @@ public class BasketController {
             description = "Create new basket by member",
             produces = MediaType.APPLICATION_JSON_VALUE,
             consumes = MediaType.APPLICATION_JSON_VALUE)
+    @ApiErrors(apierrors = {
+            @ApiError(code = BAD_REQUEST_STATUS_CODE, description = "Bad request parameter"),
+            @ApiError(code = "415", description = "Unsupported media type"),
+            @ApiError(code = "422", description = "Request element does not correct"),
+            @ApiError(code = "404", description = "Element doesn't exist"),
+            @ApiError(code = "500", description = "Unspecified server error")
+    })
     @RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity create(@ApiBodyObject @RequestBody @Valid BasketTransport basketTransport) {
+    public @ApiResponseObject ResponseEntity create(@ApiBodyObject @RequestBody @Valid BasketTransport basketTransport) {
 
         BasketModel basketModel = new BasketRequestMapper().map(basketTransport);
         BasketModel saveBasket = basketService.save(basketModel);
@@ -55,7 +68,7 @@ public class BasketController {
                 basketProductService.save(basketProductModel);
             });
         }
-        return new ResponseEntity<>(new BasketResponseMapper().map(saveBasket), HttpStatus.CREATED);
+        return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @ApiMethod(
@@ -63,13 +76,22 @@ public class BasketController {
             description = "Edit Basket detail",
             produces = MediaType.APPLICATION_JSON_VALUE,
             consumes = MediaType.APPLICATION_JSON_VALUE)
+    @ApiErrors(apierrors = {
+            @ApiError(code = BAD_REQUEST_STATUS_CODE, description = "Bad request parameter"),
+            @ApiError(code = "415", description = "Unsupported media type"),
+            @ApiError(code = "422", description = "Request element does not correct"),
+            @ApiError(code = "404", description = "Element doesn't exist"),
+            @ApiError(code = "500", description = "Unspecified server error")
+    })
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity update(@ApiBodyObject @RequestBody @Valid BasketTransport basketTransport,
+    public @ApiResponseObject ResponseEntity update(@ApiBodyObject @RequestBody @Valid BasketTransport basketTransport,
                                  @PathVariable @ApiPathParam(name = "id") Long id) {
 
         BasketModel model = basketService.findById(id);
         if(model == null){
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            LOGGER.error("Basket not found by id {}", id);
+//            throw new ResourceNotFoundException("Basket not found by id " + id);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
         BasketModel basketModel = new BasketRequestMapper().map(basketTransport);
@@ -93,18 +115,25 @@ public class BasketController {
             });
         }
 
-        return new ResponseEntity<>(new BasketResponseMapper().map(updateBasket), HttpStatus.ACCEPTED);
+        return new ResponseEntity<>(HttpStatus.ACCEPTED);
     }
 
     @ApiMethod(
             path = "/v1/baskets/{id}", verb = ApiVerb.GET,
             description = "Retrieve basket by id",
             produces = MediaType.APPLICATION_JSON_VALUE)
+    @ApiErrors(apierrors = {
+            @ApiError(code = BAD_REQUEST_STATUS_CODE, description = "Bad request parameter"),
+            @ApiError(code = NOT_FOUND_STATUS_CODE, description = "Element doesn't exist"),
+            @ApiError(code = SERVER_ERROR_STATUS_CODE, description = "Unspecified server error")
+    })
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    public ResponseEntity findById(@PathVariable @ApiPathParam(name = "id") Long id) {
-
+    public @ApiResponseObject ResponseEntity findById(@PathVariable @ApiPathParam(name = "id") Long id) {
+        LOGGER.info("Get basket by id : {}", id);
         BasketModel model = basketService.findById(id);
         if(model == null){
+            LOGGER.error("Basket not found by id {}", id);
+//            throw new ResourceNotFoundException("Basket not found by id " + id);
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
@@ -119,8 +148,13 @@ public class BasketController {
             path = "/v1/baskets", verb = ApiVerb.GET,
             description = "Retrieve all basket list",
             produces = MediaType.APPLICATION_JSON_VALUE)
+    @ApiErrors(apierrors = {
+            @ApiError(code = BAD_REQUEST_STATUS_CODE, description = "Bad request parameter"),
+            @ApiError(code = NOT_FOUND_STATUS_CODE, description = "Element doesn't exist"),
+            @ApiError(code = SERVER_ERROR_STATUS_CODE, description = "Unspecified server error")
+    })
     @RequestMapping(method = RequestMethod.GET)
-    public ResponseEntity listAll() {
+    public @ApiResponseObject ResponseEntity listAll() {
         List<BasketModel> listAll = basketService.listAll();
         List<BasketTransport> transports = new BasketResponseMapper().map(listAll);
         transports.stream().forEach(each -> {
@@ -140,8 +174,21 @@ public class BasketController {
             path = "/v1/baskets/{id}", verb = ApiVerb.DELETE,
             description = "Delete basket by id",
             produces = MediaType.APPLICATION_JSON_VALUE)
+    @ApiErrors(apierrors = {
+            @ApiError(code = "400", description = "Bad request parameter"),
+            @ApiError(code = "404", description = "Element doesn't exist"),
+            @ApiError(code = "415", description = "Unsupported media type"),
+            @ApiError(code = "422", description = "Request element does not correct"),
+            @ApiError(code = "500", description = "Unspecified server error")
+    })
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-    public ResponseEntity delete(@PathVariable @ApiPathParam(name = "id") Integer id) {
+    public @ApiResponseObject ResponseEntity delete(@PathVariable @ApiPathParam(name = "id") Integer id) {
+        BasketModel basketModel = basketService.findById(id.longValue());
+        if(basketModel == null) {
+            LOGGER.error("Basket not found by id {}", id);
+//            throw new ResourceNotFoundException("Basket not found by id " + id);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         basketService.delete(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
